UIをきれいにする

+++ src/components/TodoList.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
     <v-toolbar dark color="indigo">
-          <v-toolbar-title class="white--text">Todo APP</v-toolbar-title>
+          <v-toolbar-title class="white--text">Todo LIST</v-toolbar-title>
     </v-toolbar>
     <v-list>
         <template v-for="(todo, i) in todos">
@@ -12,9 +12,7 @@
                     {{ todo }}
                 </v-list-tile-content>
                 <v-list-tile-action>
-                    <v-btn flat icon v-on:click="deleteTodo(i)">
-                        <v-icon>DELETE</v-icon>
-                    </v-btn>
+                    <v-btn id="deleteCircleBtn" flat icon v-on:click="deleteTodo(i)">Delete</v-btn>
                 </v-list-tile-action>
             </v-list-tile>
             <v-divider v-bind:key="`second-${i}`"></v-divider>
@@ -23,7 +21,7 @@
 
     <div class="bottom-right">
         <v-btn fab color="indigo" v-on:click="addTodo">
-            <v-icon color="white">ADD</v-icon>
+            <v-icon id="addBtn" color="white">ADD</v-icon>
         </v-btn>
     </div>
   </div>
@@ -41,8 +39,13 @@ export default {
     },
     methods: {
       deleteTodo(i){
+        let res = confirm('本当に消しますか？？')
+        if(res == true){
         this.todos.splice(i, 1);
         localStorage.setItem('todos', JSON.stringify(this.todos));
+        } else {
+            alert("消すのをやめましたー！！")
+        }
       },
       addTodo(){
           this.$router.push('/todos/add');
@@ -56,5 +59,16 @@ export default {
     position: fixed;
     bottom: 0px;
     right: 0px;
+}
+#deleteCircleBtn{
+    /* padding-right: 60px; */
+    width: 140px;
+    color:red;
+}
+v-btn__contetnt {
+    text-align: center;
+}
+#addBtn {
+    font-size: 20px;
 }
 </style>
